really it should be the spellbook doing all that
